+++ TutorOnline/src/main/java/com/pmi/tutor/domain/Role.java
@@ -15,17 +15,19 @@ public class Role {
 	@Id
 	@GeneratedValue
 	@Column(name = "id")
-	private Integer id;
+	private Long id;
 	
 	@Enumerated(EnumType.STRING)
 	@Column(name = "name")
 	private RoleEnum name;
 	
-	public Integer getId() {
+	
+
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
